backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
Pillow
requests
joblib
numpy
pandas
scikit-learn
matplotlib
